+++ R/annotate_axis_ticks.R
@@ -1,157 +1,3 @@
-#' Annotated axis line segment
-#'
-#' @description Replace axis line with an annotated segment.
-#'
-#' @param position The position of the axis. One of "bottom", "top", "left", or "right".
-#' @param ... Extra parameters passed to `ggplot2::annotate("segment", ...)`.
-#' @param colour The colour of the annotated segment. Inherits from the current theme axis.line etc.
-#' @param linewidth The linewidth of the annotated segment. Inherits from the current theme axis.line etc.
-#'
-#' @return A list of a annotate layer and theme elements.
-#' @export
-#'
-annotate_axis_line <- function(position, ..., colour = NULL, linewidth = NULL) {
-  rlang::inform(
-    "Please use this function with ggplot2::coord_cartesian(clip = 'off')"
-  )
-
-  if (!position %in% c("bottom", "top", "left", "right")) {
-    rlang::abort("position must be one of 'bottom', 'top', 'left', or 'right'")
-  }
-
-  # Get current theme
-  current_theme <- ggplot2::get_theme()
-
-  # Determine if this is x-axis or y-axis
-  is_x_axis <- position %in% c("bottom", "top")
-
-  # Extract theme properties for axis line
-  if (rlang::is_null(colour)) {
-    if (is_x_axis) {
-      colour <- if (position == "bottom") {
-        current_theme$axis.line.x.bottom$colour %||%
-          current_theme$axis.line.x$colour %||%
-          current_theme$axis.line$colour %||%
-          "#121B24FF"
-      } else {
-        current_theme$axis.line.x.top$colour %||%
-          current_theme$axis.line.x$colour %||%
-          current_theme$axis.line$colour %||%
-          "#121B24FF"
-      }
-    } else {
-      colour <- if (position == "left") {
-        current_theme$axis.line.y.left$colour %||%
-          current_theme$axis.line.y$colour %||%
-          current_theme$axis.line$colour %||%
-          "#121B24FF"
-      } else {
-        current_theme$axis.line.y.right$colour %||%
-          current_theme$axis.line.y$colour %||%
-          current_theme$axis.line$colour %||%
-          "#121B24FF"
-      }
-    }
-  }
-
-  if (rlang::is_null(linewidth)) {
-    if (is_x_axis) {
-      linewidth <- if (position == "bottom") {
-        current_theme$axis.line.x.bottom$linewidth %||%
-          current_theme$axis.line.x$linewidth %||%
-          current_theme$axis.line$linewidth %||%
-          0.5
-      } else {
-        current_theme$axis.line.x.top$linewidth %||%
-          current_theme$axis.line.x$linewidth %||%
-          current_theme$axis.line$linewidth %||%
-          0.5
-      }
-    } else {
-      linewidth <- if (position == "left") {
-        current_theme$axis.line.y.left$linewidth %||%
-          current_theme$axis.line.y$linewidth %||%
-          current_theme$axis.line$linewidth %||%
-          0.5
-      } else {
-        current_theme$axis.line.y.right$linewidth %||%
-          current_theme$axis.line.y$linewidth %||%
-          current_theme$axis.line$linewidth %||%
-          0.5
-      }
-    }
-  }
-
-  # Create appropriate segment and theme based on position
-  if (position == "bottom") {
-    stamp <- rlang::list2(
-      ggplot2::annotate(
-        "segment",
-        x = I(-Inf),
-        xend = I(Inf),
-        y = I(-Inf),
-        yend = I(-Inf),
-        colour = colour,
-        linewidth = linewidth,
-        ...
-      ),
-      ggplot2::theme(
-        axis.line.x.bottom = ggplot2::element_line(colour = "transparent")
-      )
-    )
-  } else if (position == "top") {
-    stamp <- rlang::list2(
-      ggplot2::annotate(
-        "segment",
-        x = I(-Inf),
-        xend = I(Inf),
-        y = I(Inf),
-        yend = I(Inf),
-        colour = colour,
-        linewidth = linewidth,
-        ...
-      ),
-      ggplot2::theme(
-        axis.line.x.top = ggplot2::element_line(colour = "transparent")
-      )
-    )
-  } else if (position == "left") {
-    stamp <- rlang::list2(
-      ggplot2::annotate(
-        "segment",
-        x = I(-Inf),
-        xend = I(-Inf),
-        y = I(-Inf),
-        yend = I(Inf),
-        colour = colour,
-        linewidth = linewidth,
-        ...
-      ),
-      ggplot2::theme(
-        axis.line.y.left = ggplot2::element_line(colour = "transparent")
-      )
-    )
-  } else if (position == "right") {
-    stamp <- rlang::list2(
-      ggplot2::annotate(
-        "segment",
-        x = I(Inf),
-        xend = I(Inf),
-        y = I(-Inf),
-        yend = I(Inf),
-        colour = colour,
-        linewidth = linewidth,
-        ...
-      ),
-      ggplot2::theme(
-        axis.line.y.right = ggplot2::element_line(colour = "transparent")
-      )
-    )
-  }
-
-  return(stamp)
-}
-
 #' Annotated axis ticks segments
 #'
 #' @description Replace axis ticks with annotated segments using absolute measurements.
@@ -166,6 +12,37 @@ annotate_axis_line <- function(position, ..., colour = NULL, linewidth = NULL) {
 #'
 #' @return A list of a annotate layer and theme elements.
 #' @export
+#'
+#' @examples
+#' library(dplyr)
+#' library(stringr)
+#' library(palmerpenguins)
+#'
+#' set_blanket(
+#'   theme = theme_lighter(
+#'     panel_heights = rep(unit(50, "mm"), times = 100),
+#'     panel_widths = rep(unit(75, "mm"), times = 100),
+#'   ),
+#' )
+#'
+#' penguins |>
+#'   tidyr::drop_na(sex) |>
+#'   mutate(across(sex, \(x) str_to_sentence(x))) |>
+#'   gg_histogram(
+#'     x = flipper_length_mm,
+#'     col = sex,
+#'   ) +
+#'   annotate_axis_ticks(position = "bottom", breaks = c(seq(170, 230, 10), 195))
+#'
+#' penguins |>
+#'   tidyr::drop_na(sex) |>
+#'   mutate(across(sex, \(x) str_to_sentence(x))) |>
+#'   gg_bar(
+#'     x = island,
+#'     col = sex,
+#'     width = 0.75,
+#'   ) +
+#'   annotate_axis_ticks(position = "bottom", breaks = c(0.5, 1.5, 2.5, 3.5))
 #'
 annotate_axis_ticks <- function(
     position,