+++ tests/testthat/test-col_palette.R
@@ -52,6 +52,9 @@ test_that(test_name, {
   vdiffr::expect_doppelganger(test_name, p)
 })
 
+set_blanket(
+  col_palette_c = scales::pal_viridis(option = "A"),
+)
 
 test_name <- "4"
 
@@ -62,7 +65,6 @@ test_that(test_name, {
       y = eruptions,
       z = density,
       bins = 8,
-      col_palette = scales::pal_viridis(option = "A"),
     )
 
   vdiffr::expect_doppelganger(test_name, p)
@@ -78,11 +80,14 @@ test_that(test_name, {
       x = flipper_length_mm,
       y = species,
       col = bill_length_mm,
-      col_palette = scales::pal_viridis(option = "A"),
     )
 
   vdiffr::expect_doppelganger(test_name, p)
 })
+
+set_blanket(
+  col_palette_d = scales::pal_viridis(option = "A"),
+)
 
 test_name <- "6"
 
@@ -94,11 +99,12 @@ test_that(test_name, {
       x = flipper_length_mm,
       y = species,
       col = species,
-      col_palette = scales::pal_viridis(option = "A")
     )
 
   vdiffr::expect_doppelganger(test_name, p)
 })
+
+set_blanket()
 
 test_name <- "7"
 
@@ -110,11 +116,16 @@ test_that(test_name, {
       x = flipper_length_mm,
       y = species,
       col = bill_length_mm,
-      col_palette = c("#000004FF", "#07071DFF", "#160F3BFF", "#29115AFF", "#400F73FF",
+      colour_palette = c("#000004FF", "#07071DFF", "#160F3BFF", "#29115AFF", "#400F73FF",
                       "#56147DFF", "#6B1D81FF", "#802582FF",
                       "#952C80FF", "#AB337CFF", "#C03A76FF", "#D6456CFF", "#E85362FF",
                       "#F4685CFF", "#FA815FFF", "#FD9A6AFF",
-                      "#FEB37BFF", "#FECC8FFF", "#FDE4A6FF", "#FCFDBFFF")
+                      "#FEB37BFF", "#FECC8FFF", "#FDE4A6FF", "#FCFDBFFF"),
+      fill_palette = c("#000004FF", "#07071DFF", "#160F3BFF", "#29115AFF", "#400F73FF",
+                         "#56147DFF", "#6B1D81FF", "#802582FF",
+                         "#952C80FF", "#AB337CFF", "#C03A76FF", "#D6456CFF", "#E85362FF",
+                         "#F4685CFF", "#FA815FFF", "#FD9A6AFF",
+                         "#FEB37BFF", "#FECC8FFF", "#FDE4A6FF", "#FCFDBFFF")
     )
 
   vdiffr::expect_doppelganger(test_name, p)
@@ -130,7 +141,7 @@ test_that(test_name, {
       x = flipper_length_mm,
       y = species,
       col = species,
-      col_palette = c("red", "blue", "green")
+      colour_palette = c("red", "blue", "green")
     )
 
   vdiffr::expect_doppelganger(test_name, p)
@@ -146,7 +157,7 @@ test_that(test_name, {
       x = flipper_length_mm,
       y = species,
       col = species,
-      col_palette = c(
+      colour_palette = c(
         "Gentoo" = "red",
         "Adelie" = "blue",
         "Chinstrap" = "green"
@@ -155,6 +166,8 @@ test_that(test_name, {
 
   vdiffr::expect_doppelganger(test_name, p)
 })
+
+set_blanket()
 
 test_name <- "10"